refactor(core): centralize model resolution
