+++ packages/core/src/generate.ts
@@ -1,16 +1,11 @@
 export * as Generate from "./generate"
 
 import { LLM, LLMClient, LLMError } from "@opencode-ai/ai"
-import { Npm } from "@opencode-ai/util/npm"
 import { Context, Effect, Layer, Schema } from "effect"
-import { AISDK } from "./aisdk"
-import { Catalog } from "./catalog"
 import { makeLocationNode } from "@opencode-ai/util/effect/app-node"
 import { llmClient } from "./effect/app-node-platform"
-import { Integration } from "./integration"
+import { ModelResolver } from "./model-resolver"
 import { ModelV2 } from "./model"
-import { ProviderV2 } from "./provider"
-import { SessionRunnerModel } from "./session/runner/model"
 
 export interface TextInput {
   readonly prompt: string
@@ -22,10 +17,10 @@ export class ModelSelectionError extends Schema.TaggedErrorClass<ModelSelectionE
   { message: Schema.String },
 ) {}
 
-export class UnavailableError extends Schema.TaggedErrorClass<UnavailableError>()(
-  "Generate.UnavailableError",
-  { message: Schema.String, service: Schema.optional(Schema.String) },
-) {}
+export class UnavailableError extends Schema.TaggedErrorClass<UnavailableError>()("Generate.UnavailableError", {
+  message: Schema.String,
+  service: Schema.optional(Schema.String),
+}) {}
 
 export type Error = ModelSelectionError | UnavailableError
 
@@ -38,61 +33,34 @@ export class Service extends Context.Service<Service, Interface>()("@opencode/v2
 export const layer = Layer.effect(
   Service,
   Effect.gen(function* () {
-    const catalog = yield* Catalog.Service
-    const integrations = yield* Integration.Service
     const llm = yield* LLMClient.Service
-    const npm = yield* Npm.Service
-    const aisdk = yield* AISDK.Service
-
-    const selectModel = Effect.fn("Generate.selectModel")(function* (requested?: ModelV2.Ref) {
-      const selected = requested
-        ? yield* catalog.model.get(requested.providerID, requested.id)
-        : yield* catalog.model.default().pipe(
-            Effect.flatMap((model) =>
-              model && SessionRunnerModel.supported(model)
-                ? Effect.succeed(model)
-                : Effect.map(catalog.model.available(), (models) => models.find(SessionRunnerModel.supported)),
-            ),
-          )
-      if (!selected)
-        return yield* new ModelSelectionError({
-          message: requested
-            ? `Model unavailable: ${requested.providerID}/${requested.id}`
-            : "No model specified and no supported model is available",
-        })
-      return yield* SessionRunnerModel.withVariant(selected, requested?.variant).pipe(
-        Effect.mapError(
-          () =>
-            new ModelSelectionError({
-              message: `Variant unavailable for ${selected.providerID}/${selected.id}: ${requested?.variant}`,
-            }),
-        ),
-      )
-    })
+    const resolver = yield* ModelResolver.Service
 
     const runText = Effect.fn("Generate.text")(function* (input: TextInput) {
-      const selected = yield* selectModel(input.model)
-      const provider = yield* catalog.provider.get(selected.providerID)
-      const connection = yield* integrations.connection.active(
-        provider?.integrationID ?? Integration.ID.make(selected.providerID),
+      const resolved = yield* resolver.resolve(input.model).pipe(
+        Effect.catchTags({
+          "SessionRunnerModel.VariantUnavailableError": (error) =>
+            input.model
+              ? new ModelSelectionError({ message: error.message })
+              : new UnavailableError({ message: error.message, service: error.providerID }),
+          "SessionRunnerModel.UnsupportedPackageError": (error) =>
+            input.model
+              ? new ModelSelectionError({ message: error.message })
+              : new UnavailableError({ message: error.message, service: error.providerID }),
+        }),
       )
-      const credential = connection ? yield* integrations.connection.resolve(connection) : undefined
-      const model = yield* SessionRunnerModel.fromCatalogModel(selected, credential, {
-        loadPackage: (specifier) => ProviderV2.loadPackage(specifier, npm),
-        loadAISDK: (model) => aisdk.model(model),
-      }).pipe(
-        Effect.mapError((error) =>
-          input.model
-            ? new ModelSelectionError({ message: error.message })
-            : new UnavailableError({ message: error.message, service: selected.providerID }),
-        ),
-      )
-      const response = yield* llm.generate(LLM.request({ model, prompt: input.prompt })).pipe(
+      if (!resolved)
+        return yield* new ModelSelectionError({
+          message: input.model
+            ? `Model unavailable: ${input.model.providerID}/${input.model.id}`
+            : "No model specified and no supported model is available",
+        })
+      const response = yield* llm.generate(LLM.request({ model: resolved.model, prompt: input.prompt })).pipe(
         Effect.mapError(
           (error: LLMError) =>
             new UnavailableError({
               message: error.message,
-              service: selected.providerID,
+              service: resolved.ref.providerID,
             }),
         ),
       )
@@ -117,5 +85,5 @@ export const layer = Layer.effect(
 export const node = makeLocationNode({
   service: Service,
   layer,
-  deps: [Catalog.node, Integration.node, llmClient, Npm.node, AISDK.node],
+  deps: [ModelResolver.node, llmClient],
 })
