+++ packages/core/test/generate.test.ts
@@ -5,6 +5,7 @@ import { AISDK } from "@opencode-ai/core/aisdk"
 import { Catalog } from "@opencode-ai/core/catalog"
 import { Generate } from "@opencode-ai/core/generate"
 import { Integration } from "@opencode-ai/core/integration"
+import { ModelResolver } from "@opencode-ai/core/model-resolver"
 import { ModelV2 } from "@opencode-ai/core/model"
 import { ProviderV2 } from "@opencode-ai/core/provider"
 import { Npm } from "@opencode-ai/util/npm"
@@ -79,7 +80,9 @@ const client = Layer.mock(LLMClient.Service)({
     }),
 })
 
-const it = testEffect(Generate.layer.pipe(Layer.provide(Layer.mergeAll(catalog, integrations, npm, aisdk, client))))
+const resolver = ModelResolver.layer.pipe(Layer.provide(Layer.mergeAll(catalog, integrations, npm, aisdk)))
+const it = testEffect(Generate.layer.pipe(Layer.provide(Layer.merge(resolver, client))))
+const resolverIt = testEffect(resolver)
 
 it.effect("loads dynamic AI SDK models", () =>
   Effect.gen(function* () {
@@ -90,5 +93,19 @@ it.effect("loads dynamic AI SDK models", () =>
     })
 
     expect(result).toBe("OK")
+  }),
+)
+
+resolverIt.effect("resolves dynamic models with their catalog metadata", () =>
+  Effect.gen(function* () {
+    const resolver = yield* ModelResolver.Service
+    const result = yield* resolver.resolve(ModelV2.Ref.make({ providerID: selected.providerID, id: selected.id }))
+
+    expect(result).toEqual({
+      model: runtime,
+      ref: ModelV2.Ref.make({ providerID: selected.providerID, id: selected.id }),
+      capabilities: selected.capabilities,
+      cost: selected.cost,
+    })
   }),
 )
